code cleanup and adding delete functionality
