+++ clients/core/src/managementConsole/applicationAdministration/pages/ApplicationAssessment/components/ApplicationDetailsDialog/hooks/mutateAssessment.ts
@@ -1,0 +1,31 @@
+import { useToast } from '@/hooks/use-toast'
+import { ApplicationAssessment } from '@core/managementConsole/applicationAdministration/interfaces/applicationAssessment'
+import { postApplicationAssessment } from '@core/network/mutations/postApplicationAssessment'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
+import { useParams } from 'react-router-dom'
+
+export const useModifyAssessment = (coursePhaseParticipationID: string) => {
+  const { phaseId } = useParams<{ phaseId: string }>()
+  const queryClient = useQueryClient()
+  const { toast } = useToast()
+
+  return useMutation({
+    mutationFn: (applicationAssessment: ApplicationAssessment) => {
+      return postApplicationAssessment(
+        phaseId ?? 'undefined',
+        coursePhaseParticipationID,
+        applicationAssessment,
+      )
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['application_participations'] })
+    },
+    onError: () => {
+      toast({
+        title: 'Failed to Store Assessment',
+        description: 'Please try again later!',
+        variant: 'destructive',
+      })
+    },
+  })
+}
